Remove partner_semantic sandbox
